DCMAW-8229: extract function that gets registered client
