+++ backend-api/tests/api-tests/clientCredentialsGrantFlow.test.ts
@@ -6,6 +6,7 @@ import {
 import { randomUUID, UUID } from "crypto";
 import { PRIVATE_API_INSTANCE, PROXY_API_INSTANCE } from "./utils/apiInstance";
 import { AxiosInstance } from "axios";
+import { ClientDetails, getFirstRegisteredClient } from "./utils/getRegisteredClient";
 
 const getApisToTest = (): {
   apiName: string;
@@ -335,29 +336,6 @@ describe.each(apis)(
     });
   },
 );
-
-interface ClientDetails {
-  client_id: string;
-  client_secret: string;
-  redirect_uri: string;
-}
-
-async function getRegisteredClients(): Promise<ClientDetails[]> {
-  const secretsManagerClient = new SecretsManagerClient({
-    region: "eu-west-2",
-  });
-  const secretName = `${process.env.TEST_ENVIRONMENT}/clientRegistryApiTest`;
-  const command = new GetSecretValueCommand({
-    SecretId: secretName,
-  });
-  const response = await secretsManagerClient.send(command);
-  return JSON.parse(response.SecretString!);
-}
-
-async function getFirstRegisteredClient(): Promise<ClientDetails> {
-  const clientsDetails = await getRegisteredClients();
-  return clientsDetails[0];
-}
 
 function toBase64(value: string): string {
   return Buffer.from(value).toString("base64");
